refactor(runtime): extract 7 leaf tool modules from tool_runner
Pull out the no-helper, single-dispatch tools into their own files:

- `notify` — notify_owner (opaque operator ack)
- `memory` — memory_store / recall / list
- `wiki` — wiki_get / search / write
- `agent` — agent_find (agent_send/spawn/list/kill stay in mod.rs for now)
- `task` — task_post / claim / complete / list / status
- `event` — event_publish
- `goal` — goal_update

require_kernel is promoted to pub(super) so the new modules can share
the same Option<&Arc<dyn KernelHandle>> -> &Arc<dyn KernelHandle>
unwrapping behavior the original inline functions had.

Behavior-preserving: each tool is a verbatim function move with only
pub(super) visibility added; the parent dispatch sites stay unchanged
via use self::<module>::<fn> re-imports.

mod.rs: 13,451 -> 13,123 LOC (-328). cargo check + cargo clippy --lib
both clean.

Refs #3710
