+++ crates/librefang-runtime/src/tool_runner/agent.rs
@@ -1,0 +1,36 @@
+//! Inter-agent collaboration tools.
+//!
+//! `agent_send`, `agent_spawn`, `agent_list`, `agent_kill` continue to live in
+//! the parent `tool_runner/mod.rs` for now — they share large helpers with the
+//! dispatch path and will move here in a follow-up extraction.
+
+use super::require_kernel;
+use crate::kernel_handle::prelude::*;
+use std::sync::Arc;
+
+pub(super) fn tool_agent_find(
+    input: &serde_json::Value,
+    kernel: Option<&Arc<dyn KernelHandle>>,
+) -> Result<String, String> {
+    let kh = require_kernel(kernel)?;
+    let query = input["query"].as_str().ok_or("Missing 'query' parameter")?;
+    let agents = kh.find_agents(query);
+    if agents.is_empty() {
+        return Ok(format!("No agents found matching '{query}'."));
+    }
+    let result: Vec<serde_json::Value> = agents
+        .iter()
+        .map(|a| {
+            serde_json::json!({
+                "id": a.id,
+                "name": a.name,
+                "state": a.state,
+                "description": a.description,
+                "tags": a.tags,
+                "tools": a.tools,
+                "model": format!("{}:{}", a.model_provider, a.model_name),
+            })
+        })
+        .collect();
+    serde_json::to_string_pretty(&result).map_err(|e| format!("Serialize error: {e}"))
+}
